+++ src/main/java/io/hops/metadata/hdfs/dal/CorruptReplicaDataAccess.java
@@ -31,11 +31,11 @@ public interface CorruptReplicaDataAccess<T> extends EntityDataAccess {
 
   List<T> findAll() throws StorageException;
 
-  List<T> findByBlockId(long blockId, int inodeId) throws StorageException;
+  List<T> findByBlockId(long blockId, long inodeId) throws StorageException;
   
-  List<T> findByINodeId(int inodeId) throws StorageException;
+  List<T> findByINodeId(long inodeId) throws StorageException;
   
-  List<T> findByINodeIds(int[] inodeIds) throws StorageException;
+  List<T> findByINodeIds(long[] inodeIds) throws StorageException;
 
   void prepare(Collection<T> removed, Collection<T> newed) throws StorageException;
 }
