+++ src/main/java/io/hops/metadata/hdfs/dal/INodeAttributesDataAccess.java
@@ -24,7 +24,7 @@ import java.util.List;
 
 public interface INodeAttributesDataAccess<T> extends EntityDataAccess {
 
-  T findAttributesByPk(Integer inodeId) throws StorageException;
+  T findAttributesByPk(Long inodeId) throws StorageException;
   
   Collection<T> findAttributesByPkList(List<INodeCandidatePrimaryKey> inodePks)
       throws StorageException;
